+++ test/license_test.dart
@@ -6,19 +6,27 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:pana/src/license.dart';
-import 'package:pana/src/model.dart';
 import 'package:test/test.dart';
 
 void main() {
-  Future expectFile(String path, LicenseFile expected) async {
+  Future expectFile(String path, expected) async {
     final relativePath = path.substring('test/licenses/'.length);
-    expect(await detectLicenseInFile(File(path), relativePath: relativePath),
+    final licenses =
+        await detectLicenseInFile(File(path), relativePath: relativePath);
+    expect(
+        licenses
+            .map((e) => {'spdx': e.spdx, 'confidence': e.confidence})
+            .toList(),
         expected);
   }
 
   test('bad encoding', () async {
-    await expectFile('test/licenses/bad_encoding.txt',
-        LicenseFile('bad_encoding.txt', 'Zlib'));
+    await expectFile('test/licenses/bad_encoding.txt', [
+      {
+        'spdx': 'Zlib',
+        'confidence': closeTo(0.98, 0.01),
+      },
+    ]);
   });
 
   group('AGPL', () {
@@ -26,9 +34,13 @@ void main() {
       expect(
           await detectLicenseInContent('GNU AFFERO GENERAL PUBLIC LICENSE',
               relativePath: 'LICENSE'),
-          null);
-      await expectFile(
-          'test/licenses/agpl_v3.txt', LicenseFile('agpl_v3.txt', 'AGPL-3.0'));
+          []);
+      await expectFile('test/licenses/agpl_v3.txt', [
+        {
+          'spdx': 'AGPL-3.0',
+          'confidence': closeTo(0.99, 01),
+        },
+      ]);
     });
   });
 
@@ -38,22 +50,39 @@ void main() {
           await detectLicenseInContent(
               '   Apache License\n     Version 2.0, January 2004\n',
               relativePath: 'LICENSE'),
-          null);
-      await expectFile('test/licenses/apache_v2.txt',
-          LicenseFile('apache_v2.txt', 'Apache-2.0'));
+          []);
+      await expectFile('test/licenses/apache_v2.txt', [
+        {'spdx': 'Apache-2.0', 'confidence': 1.0}
+      ]);
     });
   });
 
   group('BSD', () {
     test('explicit', () async {
-      await expectFile('test/licenses/bsd_2_clause.txt',
-          LicenseFile('bsd_2_clause.txt', 'BSD-2-Clause'));
-      await expectFile('test/licenses/bsd_2_clause_in_comments.txt',
-          LicenseFile('bsd_2_clause_in_comments.txt', 'BSD-2-Clause'));
-      await expectFile('test/licenses/bsd_3_clause.txt',
-          LicenseFile('bsd_3_clause.txt', 'unknown'));
-      await expectFile('test/licenses/bsd_revised.txt',
-          LicenseFile('bsd_revised.txt', 'BSD-3-Clause'));
+      await expectFile('test/licenses/bsd_2_clause.txt', [
+        {
+          'spdx': 'BSD-2-Clause',
+          'confidence': closeTo(0.98, 0.01),
+        },
+      ]);
+      await expectFile('test/licenses/bsd_2_clause_in_comments.txt', [
+        {
+          'spdx': 'BSD-2-Clause',
+          'confidence': closeTo(0.98, 0.01),
+        },
+      ]);
+      await expectFile('test/licenses/bsd_3_clause.txt', [
+        {
+          'spdx': 'unknown',
+          'confidence': 0.0,
+        },
+      ]);
+      await expectFile('test/licenses/bsd_revised.txt', [
+        {
+          'spdx': 'BSD-3-Clause',
+          'confidence': closeTo(0.98, 0.01),
+        },
+      ]);
     });
   });
 
@@ -63,13 +92,17 @@ void main() {
           await detectLicenseInContent(
               ['GNU GENERAL PUBLIC LICENSE', 'Version 2, June 1991'].join('\n'),
               relativePath: 'LICENSE'),
-          null);
+          []);
       expect(
           await detectLicenseInContent(['GNU GPL Version 2'].join('\n'),
               relativePath: 'LICENSE'),
-          null);
-      await expectFile(
-          'test/licenses/gpl_v3.txt', LicenseFile('gpl_v3.txt', 'GPL-3.0'));
+          []);
+      await expectFile('test/licenses/gpl_v3.txt', [
+        {
+          'spdx': 'GPL-3.0',
+          'confidence': closeTo(0.99, 0.01),
+        },
+      ]);
     });
   });
 
@@ -79,13 +112,10 @@ void main() {
           await detectLicenseInContent(
               '\nGNU LESSER GENERAL PUBLIC LICENSE\n    Version 3, 29 June 2007',
               relativePath: 'LICENSE'),
-          null);
-      await expectFile(
-          'test/licenses/lgpl_v3.txt',
-          LicenseFile(
-            'lgpl_v3.txt',
-            'LGPL-3.0',
-          ));
+          []);
+      await expectFile('test/licenses/lgpl_v3.txt', [
+        {'spdx': 'LGPL-3.0', 'confidence': 1.0}
+      ]);
     });
   });
 
@@ -94,14 +124,24 @@ void main() {
       expect(
           await detectLicenseInContent('\n\n   The MIT license\n\n blah...',
               relativePath: 'LICENSE'),
-          null);
+          []);
       expect(
           await detectLicenseInContent('MIT license\n\n blah...',
               relativePath: 'LICENSE'),
-          null);
-      await expectFile('test/licenses/mit.txt', LicenseFile('mit.txt', 'MIT'));
-      await expectFile('test/licenses/mit_without_mit.txt',
-          LicenseFile('mit_without_mit.txt', 'MIT'));
+          []);
+
+      await expectFile('test/licenses/mit.txt', [
+        {
+          'spdx': 'MIT',
+          'confidence': closeTo(0.96, 0.01),
+        },
+      ]);
+      await expectFile('test/licenses/mit_without_mit.txt', [
+        {
+          'spdx': 'MIT',
+          'confidence': closeTo(0.96, 0.01),
+        },
+      ]);
     });
   });
 
@@ -111,9 +151,10 @@ void main() {
           await detectLicenseInContent(
               '\n\n   Mozilla Public License Version 2.0\n\n blah...',
               relativePath: 'LICENSE'),
-          null);
-      await expectFile(
-          'test/licenses/mpl_v2.txt', LicenseFile('mpl_v2.txt', 'MPL-2.0'));
+          []);
+      await expectFile('test/licenses/mpl_v2.txt', [
+        {'spdx': 'MPL-2.0', 'confidence': 1.0}
+      ]);
     });
   });
 
@@ -123,26 +164,35 @@ void main() {
           await detectLicenseInContent(
               '\n\n   This is free and unencumbered software released into the public domain.\n',
               relativePath: 'LICENSE'),
-          null);
-      await expectFile('test/licenses/unlicense.txt',
-          LicenseFile('unlicense.txt', 'Unlicense'));
+          []);
+      await expectFile('test/licenses/unlicense.txt', [
+        {'spdx': 'Unlicense', 'confidence': 1.0}
+      ]);
     });
   });
 
   group('unknown', () {
     test('empty content', () async {
-      expect(await detectLicenseInContent('', relativePath: 'LICENSE'), isNull);
+      expect(await detectLicenseInContent('', relativePath: 'LICENSE'), []);
     });
   });
 
   group('Directory scans', () {
     test('detect pana LICENSE', () async {
-      expect(await detectLicenseInDir('.'),
-          LicenseFile('LICENSE', 'BSD-3-Clause'));
+      final licenses = await detectLicenseInDir('.');
+      expect(licenses.map((e) => e.toJson()).toList(), [
+        {
+          'path': 'LICENSE',
+          'spdx': 'BSD-3-Clause',
+          'confidence': closeTo(0.95, 0.01),
+          'start': isPositive,
+          'end': isPositive,
+        }
+      ]);
     });
 
     test('no license files', () async {
-      expect(await detectLicenseInDir('lib/src/'), null);
+      expect(await detectLicenseInDir('lib/src/'), []);
     });
   });
 }